+++ lightly_studio_view/src/lib/components/AnnotationDetails/AnnotationDetailsPanel/AnnotationDetailsPanel.svelte
@@ -5,15 +5,23 @@
     import SegmentTags from '../../SegmentTags/SegmentTags.svelte';
     import { type AnnotationView } from '$lib/api/lightly_studio_local';
     import type { Snippet } from 'svelte';
+    import { page } from '$app/state';
+    import { Button } from '$lib/components/ui';
+    import { useDeleteAnnotation } from '$lib/hooks/useDeleteAnnotation/useDeleteAnnotation';
+    import { toast } from 'svelte-sonner';
+    import { useAnnotationDeleteNavigation } from '$lib/hooks/useAnnotationDeleteNavigation/useAnnotationDeleteNavigation';
+    import DeleteAnnotationPopUp from '$lib/components/DeleteAnnotationPopUp/DeleteAnnotationPopUp.svelte';
 
     const {
         annotation,
         onUpdate,
-        sampleDetails
+        children,
+        collectionId
     }: {
         annotation: AnnotationView;
         onUpdate: () => void;
-        sampleDetails: Snippet;
+        children: Snippet;
+        collectionId: string;
     } = $props();
     const { removeTagFromAnnotation } = useRemoveTagFromAnnotation();
 
@@ -22,6 +30,37 @@
     const onRemoveTag = async (tagId: string) => {
         await removeTagFromAnnotation(annotation.sample_id, tagId);
     };
+
+    const { isEditingMode } = page.data.globalStorage;
+
+    const { deleteAnnotation } = useDeleteAnnotation({
+        collectionId
+    });
+
+    const handleDeleteAnnotation = async () => {
+        try {
+            await deleteAnnotation(annotation.sample_id);
+
+            toast.success('Annotation deleted successfully');
+
+            gotoNextAnnotation();
+        } catch {
+            toast.error('Failed to delete annotation. Please try again.');
+        }
+    };
+
+    const datasetId = $derived(page.params.dataset_id ?? page.data.datasetId);
+    const collectionType = $derived(page.params.collection_type ?? page.data.collectionType);
+
+    const { gotoNextAnnotation } = $derived.by(() =>
+        useAnnotationDeleteNavigation({
+            collectionId,
+            datasetId,
+            collectionType,
+            annotationIndex: page.data.annotationIndex,
+            annotationAdjacents: page.data.annotationAdjacents
+        })
+    );
 </script>
 
 <Card className="h-full">
@@ -32,7 +71,17 @@
             <SegmentTags {tags} onClick={onRemoveTag} />
             <AnnotationMetadata {annotation} {onUpdate} />
 
-            {@render sampleDetails()}
+            {@render children()}
+
+            {#if $isEditingMode}
+                <DeleteAnnotationPopUp onDelete={handleDeleteAnnotation}>
+                    <Button
+                        variant="destructive"
+                        class="w-full"
+                        data-testid="delete-annotation-trigger">Delete annotation</Button
+                    >
+                </DeleteAnnotationPopUp>
+            {/if}
         </div>
     </CardContent>
 </Card>