+++ lightly_studio_view/src/routes/datasets/[dataset_id]/[collection_type]/[collection_id]/frames/[sample_id]/+page.svelte
@@ -74,6 +74,7 @@
 
 {#if sample && sampleItem}
     <SampleDetailsPanel
+        dataTestId="video-frame-details"
         collectionId={collection_id}
         {sampleId}
         sampleURL={`${PUBLIC_VIDEOS_FRAMES_MEDIA_URL}/${sample.sample_id}`}